requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
Pillow==10.0.1

# Optional: Production deployment
gunicorn==21.2.0
python-dotenv==1.0.0 